FileGDB to SpatialDatabase
renamed FileGDB to SpatialDatabase
added deprecation warning
